+++ src/hooks/dialogs/useConfirmationDialog.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-import ConfirmationDialog from "@/components/ConfirmationDialog";
+import ConfirmationDialog from "@/components/dialogs/ConfirmationDialog";
 
 export type OpenConfirmationDialogProps = Pick<
   React.ComponentProps<typeof ConfirmationDialog>,